% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected Value.R
\name{ExpectedValue}
\alias{ExpectedValue}
\title{ExpectedValue}
\usage{
ExpectedValue(f, variable, upper_boundx, upper_boundy, upper_boundF,
  lower_boundx, lower_boundy, lower_boundF, Rep = 10000)
}
\arguments{
\item{f}{the PDF or JDF provided by the user}

\item{variable}{provided by user to suggest which variable they want to use between "X" or "XY"}

\item{upper_boundx}{The maximum value for the X Variable}

\item{upper_boundy}{The maximum value for the Y Variable}

\item{upper_boundF}{The maximum value of the probability distribution (Either a PDF or a JDF)}

\item{lower_boundx}{The minimum value for the X Variable}

\item{lower_boundy}{The minimum value for the Y variable}

\item{lower_boundF}{The minimum value of the probability distribution (Either a PDF or a JDF)}

\item{Rep}{the amount of replication times the user wants, default value is 10000}
}
\value{
An Expected Value of a f(x) or f(x,y)
}
\description{
This file aims to produce an Expected Value of a f(x) or f(x,y) given a function f
}
\examples{
Example1
f = function(x) {return(1/2/pi *(sin(x) + 1))}
ExpectedValue(f, "X", lower_boundx = 0, upper_boundx = 2* pi)

Example2
f = function(x,y) {return(x+y)}
ExpectedValue(f, variable = "XY", upper_boundx = 1, upper_boundy = 1, upper_boundF = 2,
lower_boundx = 0, lower_boundy = 0, lower_boundF = 0)

}
\author{
Youssef Younis  <email: youssef.younis@slu.edu>
}
