% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected Value.R
\name{ExpectValue}
\alias{ExpectValue}
\title{ExpectedValue}
\usage{
ExpectValue(x)
}
\arguments{
\item{Function}{g should be provided for}
}
\value{
An Expected Value of a g(x) or g(x,y)
}
\description{
This file aims to produce an Expected Value of a g(x) or g(x,y) given a function g
}
\author{
Youssef Younis  <email: youssef.younis@slu.edu>
}
