% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{Prob}
\alias{Prob}
\title{Probability Related to X or X and Y}
\usage{
Prob(f, variable, upper_boundx, upper_boundy, upper_boundF, upper_boundP,
  lower_boundx, lower_boundy, lower_boundF, lower_boundP, Rep = 10000)
}
\arguments{
\item{f}{the PDF provided by the user}

\item{variable}{provided by user to suggest which variable they want to use between "X" or "XY"}

\item{upper_boundx}{The maximum value for the X Variable}

\item{upper_boundy}{The maximum value for the Y Variable}

\item{upper_boundF}{The maximum value of the probability distribution (Either a PDF or a JDF)}

\item{upper_boundP}{The upperbound assigned by the user to calculate the probabilities}

\item{lower_boundx}{The minimum value for the X Variable}

\item{lower_boundy}{The minimum value for the Y variable}

\item{lower_boundF}{The minimum value of the probability distribution (Either a PDF or a JDF)}

\item{lower_boundP}{The lowerbound assigned by the suer to calculate the probabilities}

\item{Rep}{the amount of replication times the user wants, default value is 10000}
}
\value{
A probability related to X or X and Y
}
\description{
This file aims to produce a Probability of an X or an X and Y
}
\examples{
\dontrun {
Example1
f = function(x) {return(4*x)}
Prob(f, variable = "X", upper_boundx = .5, upper_boundP = .3, lower_boundx = 0, lower_boundP = 0, Rep = 10000)

ExampleSpeegle
f = function(x) {return(1/2/pi *(sin(x) + 1))}
Prob(f, variable = "X", upper_boundx = 2*pi, upper_boundP = 2*pi, lower_boundP = pi, lower_boundx = 0, Rep = 10000)

Example3
f = function(x,y) {return(x+y)}
Prob(f, variable = "XY", upper_boundx = 1, upper_boundy = 1, upper_boundF = 2, upper_boundP = 1, lower_boundx = 0, lower_boundy = 0, lower_boundF = 0, Rep = 10000)
}

}
\author{
Youssef Younis  <email: youssef.younis@slu.edu>
}
