% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test1.R
\name{SampleJDF}
\alias{SampleJDF}
\title{Test}
\usage{
SampleJDF(lower_boundx, lower_boundy, upper_boundx, upper_boundy,
  lower_boundJDF, upper_boundJDF, f = f, Rep = 10000)
}
\arguments{
\item{lower_boundx}{provided by user to calculate minimum boundary of "X"}

\item{lower_boundy}{provided by user to calculate minimum boundary of "Y"}

\item{upper_boundx}{provided by user to calculate maximum boundary of "X"}

\item{upper_boundy}{provided by user to calculate maximum boundary of "Y"}

\item{lower_boundJDF}{provided by user to calculate minimum boundary of "JDF"}

\item{upper_boundJDF}{provided by user to calculate maximum boundary of "JDF"}

\item{Joint}{Density Function should be provided by user}
}
\value{
Specific Rejection Sample from provided JDF
}
\description{
This file aims to produce a sample from a given Joint Density Function
}
\examples{
gx = function(x,y){
return (x+y)}
SampleJDF(0, 0, 1, 1, 0, 2, gx)
}
\author{
Youssef Younis  <email: youssef.younis@slu.edu>
}
