% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleX.R
\name{sampleX}
\alias{sampleX}
\title{SamplePDF}
\usage{
sampleX(f, lower_bound, upper_bound, interval = 0.01)
}
\arguments{
\item{f}{should be provided by user}

\item{lower_bound}{provided by user for closed specific sample}

\item{upper_bound}{provided by user for closed specific sample}

\item{interval}{Default value provided but can be altered if necessary}
}
\value{
BigSample data frame for provided PDF
}
\description{
This file aims to produce a sample from a given pdf
}
\examples{
gx <- function(x){return(4*x)}
SpecificSample <- mutate(data.frame(x = seq(0, sqrt(.5), .01)), y = gx(x))
max_c(0, sqrt(.5), y = 4*x)
ApproveReject(0, sqrt(.5), y = gx(x))
head(Replication(0, sqrt(.5), y = gx(x)))
}
\author{
Youssef Younis  <email: youssef.younis@slu.edu>
}
